inverted index working on all tests, should add more advanced tests

+++ src/utils.h
@@ -13,7 +13,7 @@ float char_to_float(char* c, uint32_t len){
     return f;
 }
 
-float char_to_uint32_t(char* c){
+uint32_t char_to_uint32_t(char* c){
     uint32_t u;
     memcpy(&u, c, sizeof(uint32_t));
     return u;
